+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2024-02-20
+### Fixed
+- fix rounding on `maxRedeem`
+- fix rounding on `maxBorrow`
+
 ## [0.3.1] - 2024-02-19
 ### Fixed
 - optimise `maxWithdraw`: do not run `getTotalCollateralAssetsWithInterest` twice